refactor duplicate codes
