--- src/main/java/br/com/openenade/api/estado/EstadoUtils.java
@@ -1,14 +1,0 @@
-package br.com.openenade.api.estado;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-
-public class EstadoUtils {
-
-    public static ResponseEntity<String> getResponseEntity(HttpStatus status, String message) {
-
-        return ResponseEntity.status(status).body("{\"Response\": \"" + message + "\"}");
-
-    }
-
-}
